+++ noodles-bgzf/CHANGELOG.md
@@ -10,6 +10,12 @@
 
   * bgzf: Raise minimum supported Rust version (MSRV) to 1.73.0.
 
+### Deprecated
+
+  * bgzf: Deprecate `AsyncReader` re-export.
+
+    Use `bgzf::r#async::Reader` instead.
+
 ## 0.34.0 - 2024-12-12
 
 ### Added